Тесты на verify

+++ internal/service/refresh.go
@@ -4,48 +4,55 @@ import (
 	"context"
 	auth "crypto_analyzer_auth_service/gen/go"
 	"crypto_analyzer_auth_service/internal/domain"
+	"crypto_analyzer_auth_service/internal/errors_my"
 	"crypto_analyzer_auth_service/internal/logger"
 	"go.uber.org/zap"
 )
 
 func (s *AuthService) Refresh(ctx context.Context, req *auth.RefreshRequest) (*auth.RefreshResponse, error) {
+	refreshToken := req.GetRefreshToken()
+	if refreshToken == "" {
+		logger.Log.Info("empty refresh token")
+		return nil, errors_my.ErrRefreshFailed
+	}
+
 	userID, err := s.Session.GetUserIDByRefreshToken(ctx, req.RefreshToken)
 	if err != nil {
 		logger.Log.Error("failed to get userID by refresh token", zap.Error(err))
-		return nil, ErrRefreshFailed
+		return nil, errors_my.ErrRefreshFailed
 	}
 
 	err = s.Session.DeleteRefreshToken(ctx, req.RefreshToken)
 	if err != nil {
 		logger.Log.Error("failed to delete old refresh token", zap.Error(err))
-		return nil, ErrRefreshFailed
+		return nil, errors_my.ErrRefreshFailed
 	}
 
 	var user *domain.User
 	user, err = s.Storage.GetUserByUserID(ctx, userID)
 	if err != nil {
 		logger.Log.Error("failed to get user data by userID", zap.Error(err))
-		return nil, ErrRefreshFailed
+		return nil, errors_my.ErrRefreshFailed
 	}
 
 	var accessToken string
 	accessToken, err = s.JWTManager.GenerateAccessToken(userID, user.Username, user.Email)
 	if err != nil {
 		logger.Log.Error("failed to generate access token", zap.Error(err))
-		return nil, ErrRefreshFailed
+		return nil, errors_my.ErrRefreshFailed
 	}
 
 	var newRefreshToken string
 	newRefreshToken, err = s.JWTManager.GenerateRefreshToken()
 	if err != nil {
 		logger.Log.Error("failed to generate new refresh token", zap.Error(err))
-		return nil, ErrRefreshFailed
+		return nil, errors_my.ErrRefreshFailed
 	}
 
 	err = s.Session.SaveRefreshToken(ctx, userID, newRefreshToken)
 	if err != nil {
 		logger.Log.Error("failed to save new refresh token", zap.Error(err))
-		return nil, ErrRefreshFailed
+		return nil, errors_my.ErrRefreshFailed
 	}
 
 	return &auth.RefreshResponse{
